refactor: move creds code into its own file
